# DBSCAN optimizations for CPU