--- a/cpp/daal/src/algorithms/k_nearest_neighbors/kdtree_knn_impl.i
+++ b/cpp/daal/src/algorithms/k_nearest_neighbors/kdtree_knn_impl.i
@@ -28,20 +28,6 @@
     #include <immintrin.h>
 #endif
 
-#if defined(_MSC_VER)
-    #define DAAL_FORCEINLINE   __forceinline
-    #define DAAL_FORCENOINLINE __declspec(noinline)
-#else
-    #define DAAL_FORCEINLINE   inline __attribute__((always_inline))
-    #define DAAL_FORCENOINLINE __attribute__((noinline))
-#endif
-
-#if defined(_MSC_VER) && (_MSC_VER < 1900)
-    #define DAAL_ALIGNAS(n) __declspec(align(n))
-#else
-    #define DAAL_ALIGNAS(n) alignas(n)
-#endif
-
 namespace daal
 {
 namespace algorithms