@SmirnovEgorRu 2020-07-06
/intelci: run