@vepifanov 2020-07-06
Looks great!
Right now, I don't have any objections on code.

Btw, it would be interesting to see performance comparison on various scenarios, that can appear during hyperparameters tuning:
- Small epsilon, when all neighborhoods are very small and excessive nested parallelism can give big overhead;
- Large espilon, when all neighborhoods are very big and data racing effects can appear (simultineous checks if point is already clustered can lead to situation, when point will be added to multiple queues in different threads).