@SmirnovEgorRu 2020-07-07
/intelci: run