@SmirnovEgorRu 2020-07-07
@vepifanov 

Eps | master, ms | this PR, ms | Speedup
-- | -- | -- | --
->0 | 2440.84 | 200.148 | 12.2
36 | 4018.27 | 690.756 | 5.82
1e10 | 27167.3 | 20293.1 | 1.34

100k rows x 10 features

When all samples are one cluster I suppose we spend most of time in memory allocation/memcopy/deallocation. Potentially it can be optimized by developing special allocator with some logic.